% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{anyDuplicated.uint64}
\alias{anyDuplicated.uint64}
\title{Generalized” more efficient shortcut for any(duplicated(.)).}
\usage{
\method{anyDuplicated}{uint64}(x, ...)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Generalized” more efficient shortcut for any(duplicated(.)).
}
