% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{AreOverlappingRegionKeys}
\alias{AreOverlappingRegionKeys}
\title{Checks whether two RegionKeys overlap.}
\usage{
AreOverlappingRegionKeys(rka, rkb)
}
\arguments{
\item{rka}{RegionKey A.}

\item{rkb}{RegionKey B.}
}
\value{
1 if the regions overlap, 0 otherwise.
}
\description{
Checks whether two RegionKeys overlap.
}
