% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{AreOverlappingRegions}
\alias{AreOverlappingRegions}
\title{Checks whether two regions overlap.}
\usage{
AreOverlappingRegions(
  a_chrom,
  a_startpos,
  a_endpos,
  b_chrom,
  b_startpos,
  b_endpos
)
}
\arguments{
\item{a_chrom}{Region A chromosome code.}

\item{a_startpos}{Region A start position.}

\item{a_endpos}{Region A end position (startpos + region length).}

\item{b_chrom}{Region B chromosome code.}

\item{b_startpos}{Region B start position.}

\item{b_endpos}{Region B end position (startpos + region length).}
}
\value{
1 if the regions overlap, 0 otherwise.
}
\description{
Checks whether two regions overlap.
}
