% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{AreOverlappingVariantKeyRegionKey}
\alias{AreOverlappingVariantKeyRegionKey}
\title{Checks whether a VariantKey and a RegionKey overlap.}
\usage{
AreOverlappingVariantKeyRegionKey(vk, rk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey code.}

\item{rk}{RegionKey code.}

\item{mc}{Memory-mapped columns object as returned by MmapNRVKfile.}
}
\value{
1 if the regions overlap, 0 otherwise.
}
\description{
Checks whether a VariantKey and a RegionKey overlap.
}
