% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex64.R
\name{as.character.hex64}
\alias{as.character.hex64}
\title{Coerce hex64 vector to decimal character}
\usage{
\method{as.character}{hex64}(x, ...)
}
\arguments{
\item{x}{hex64 vector}
}
\description{
Coerce hex64 vector to decimal character
}
