% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.character.uint64}
\alias{as.character.uint64}
\title{Coerce uint64 vector to character}
\usage{
\method{as.character}{uint64}(x, ...)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Coerce uint64 vector to character
}
