% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.hex.uint64}
\alias{as.hex.uint64}
\title{Coerce hexadecimal character vector vector to uint64.}
\usage{
as.hex.uint64(x, ...)
}
\arguments{
\item{x}{hexadecimal character vector (16 characters per item)}
}
\description{
Coerce hexadecimal character vector vector to uint64.
}
