% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex64.R
\name{as.hex64.character}
\alias{as.hex64.character}
\title{Coerce character vector to hex64}
\usage{
\method{as.hex64}{character}(x, ...)
}
\arguments{
\item{x}{character vector}
}
\description{
Coerce character vector to hex64
}
