% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex64.R
\name{as.hex64.uint64}
\alias{as.hex64.uint64}
\title{Convert uint64 vector to hex64.}
\usage{
\method{as.hex64}{uint64}(x, ...)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Convert uint64 vector to hex64.
}
