% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.uint64.character}
\alias{as.uint64.character}
\title{Coerce character vector to uint64}
\usage{
\method{as.uint64}{character}(x, ...)
}
\arguments{
\item{x}{character vector}
}
\description{
Coerce character vector to uint64
}
