% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.uint64.double}
\alias{as.uint64.double}
\title{Coerce double vector to uint64}
\usage{
\method{as.uint64}{double}(x, ...)
}
\arguments{
\item{x}{double vector}
}
\description{
Coerce double vector to uint64
}
