% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.uint64.hex}
\alias{as.uint64.hex}
\title{Coerce uint64 vector to hexadecimal character vector.}
\usage{
\method{as.uint64}{hex}(x, ...)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Coerce uint64 vector to hexadecimal character vector.
}
