% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.uint64.hex64}
\alias{as.uint64.hex64}
\title{Coerce hex64 vector to uint64}
\usage{
\method{as.uint64}{hex64}(x, ...)
}
\arguments{
\item{x}{hexadecimal character vector}
}
\description{
Coerce hex64 vector to uint64
}
