% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.uint64.integer}
\alias{as.uint64.integer}
\title{Coerce integer vector to uint64}
\usage{
\method{as.uint64}{integer}(x, ...)
}
\arguments{
\item{x}{integer vector}
}
\description{
Coerce integer vector to uint64
}
