% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.uint64.logical}
\alias{as.uint64.logical}
\title{Coerce logical vector to uint64}
\usage{
\method{as.uint64}{logical}(x, ...)
}
\arguments{
\item{x}{logical vector}
}
\description{
Coerce logical vector to uint64
}
