% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.uint64.uint64}
\alias{as.uint64.uint64}
\title{Coerce from uint64 to uint64.}
\usage{
\method{as.uint64}{uint64}(x, ...)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Coerce from uint64 to uint64.
}
