% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{c.uint64}
\alias{c.uint64}
\title{Concatenate uint64 vectors.}
\usage{
\method{c}{uint64}(..., recursive = FALSE)
}
\arguments{
\item{Two}{or more vectors coerced to uint64}
}
\description{
Concatenate uint64 vectors.
}
