% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{CheckReference}
\alias{CheckReference}
\title{Checks a reference allele against the genome reference data.}
\usage{
CheckReference(chrom, pos, ref, mf = vk.env$genoref_$MF)
}
\arguments{
\item{chrom}{Encoded Chromosome number (see encode_chrom).}

\item{pos}{Position. The reference position, with the first base having position 0.}

\item{ref}{Reference allele. String containing a sequence of nucleotide letters.}

\item{mf}{Memory-mapped file object as returned by MmapGenorefFile.}
}
\value{
Positive number in case of success, negative in case of error:
  *  0 the reference allele matches the reference genome;
  *  1 the reference allele is inconsistent with the genome reference (i.e. when contains nucleotide letters other than A, C, G and T);
  * -1 the reference allele does not match the reference genome;
  * -2 the reference allele is longer than the genome reference sequence.
}
\description{
Checks a reference allele against the genome reference data.
}
