% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{CloseVariantKey}
\alias{CloseVariantKey}
\title{Unmaps the memory mapped files.
This must be the last function called.}
\usage{
CloseVariantKey()
}
\description{
Unmaps the memory mapped files.
This must be the last function called.
}
