% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{DecodeChrom}
\alias{DecodeChrom}
\title{Decodes a chromosome numerical code into its string representation.}
\usage{
DecodeChrom(code)
}
\arguments{
\item{code}{Chromosome numerical code.}
}
\description{
Decodes a chromosome numerical code into its string representation.
}
