% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{DecodeRefAlt}
\alias{DecodeRefAlt}
\title{Decodes a 32 bit REF+ALT code if it was produced by the reversible encoding (11 or less bases in total, containing only A, C, G and T letters).}
\usage{
DecodeRefAlt(code)
}
\arguments{
\item{code}{REF+ALT code}
}
\description{
Decodes a 32 bit REF+ALT code if it was produced by the reversible encoding (11 or less bases in total, containing only A, C, G and T letters).
}
