% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{DecodeRegionKey}
\alias{DecodeRegionKey}
\title{Splits a RegionKey into its encoded components.}
\usage{
DecodeRegionKey(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Splits a RegionKey into its encoded components.
}
