% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{DecodeRegionStrand}
\alias{DecodeRegionStrand}
\title{Decodes a strand code: 0 to 0, 1 to +1, 2 to -1.}
\usage{
DecodeRegionStrand(strand)
}
\arguments{
\item{code}{Strand code.}
}
\description{
Decodes a strand code: 0 to 0, 1 to +1, 2 to -1.
}
