% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{DecodeVariantKey}
\alias{DecodeVariantKey}
\title{Splits a VariantKey into its CHROM, POS and REF+ALT components.}
\usage{
DecodeVariantKey(vk)
}
\arguments{
\item{vk}{Decoded variantkey structure.}

\item{code}{VariantKey code.}
}
\description{
Splits a VariantKey into its CHROM, POS and REF+ALT components.
}
