% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{duplicated.uint64}
\alias{duplicated.uint64}
\title{Determines which elements of a vector or data frame are duplicates of elements with smaller subscripts, and returns a logical vector indicating which elements (rows) are duplicates.}
\usage{
\method{duplicated}{uint64}(x, ...)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Determines which elements of a vector or data frame are duplicates of elements with smaller subscripts, and returns a logical vector indicating which elements (rows) are duplicates.
}
