% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{EncodeChrom}
\alias{EncodeChrom}
\title{Encodes a chromosome identifier into a numerical code.}
\usage{
EncodeChrom(chrom)
}
\arguments{
\item{chrom}{Chromosome. An identifier from the reference genome, no white-space or leading zeros permitted.}
}
\description{
Encodes a chromosome identifier into a numerical code.
}
