% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{EncodeRefAlt}
\alias{EncodeRefAlt}
\title{Encodes a REF+ALT pair into a 31 bit code.}
\usage{
EncodeRefAlt(ref, alt)
}
\arguments{
\item{ref}{Reference allele. String containing a sequence of nucleotide letters.}

\item{alt}{Alternate non-reference allele string.}
}
\description{
Encodes a REF+ALT pair into a 31 bit code.
}
