% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{EncodeRegionStrand}
\alias{EncodeRegionStrand}
\title{Encodes a strand direction: -1 to 2, 0 to 0, +1 to 1.}
\usage{
EncodeRegionStrand(strand)
}
\arguments{
\item{strand}{Strand direction (-1, 0, +1).}
}
\description{
Encodes a strand direction: -1 to 2, 0 to 0, +1 to 1.
}
