% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{EncodeStringNumID}
\alias{EncodeStringNumID}
\title{Encodes a string made of a character section, a separator and a numerical section
into a 64 bit unsigned integer. For example: "ABCDE:0001234"
Encodes up to 5 characters in uppercase, a number up to 2^27, and up to 7 zero padding digits.
Strings of 10 characters or less are encoded as by EncodeStringID().}
\usage{
EncodeStringNumID(str, sep = ":")
}
\arguments{
\item{str}{The string to encode. It supports ASCII characters from '!' to 'z'.}

\item{sep}{Separator character between string and number.}
}
\description{
Encodes a string made of a character section, a separator and a numerical section
into a 64 bit unsigned integer. For example: "ABCDE:0001234"
Encodes up to 5 characters in uppercase, a number up to 2^27, and up to 7 zero padding digits.
Strings of 10 characters or less are encoded as by EncodeStringID().
}
