% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtendRegionKey}
\alias{ExtendRegionKey}
\title{Extends a RegionKey region by a fixed amount at both ends.}
\usage{
ExtendRegionKey(rk, size)
}
\arguments{
\item{rk}{RegionKey code.}

\item{size}{Amount to extend the region.}
}
\description{
Extends a RegionKey region by a fixed amount at both ends.
}
