% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractRegionKeyChrom}
\alias{ExtractRegionKeyChrom}
\title{Extracts the CHROM code from a RegionKey.}
\usage{
ExtractRegionKeyChrom(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Extracts the CHROM code from a RegionKey.
}
