% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractRegionKeyEndPos}
\alias{ExtractRegionKeyEndPos}
\title{Extracts the END POS value from a RegionKey.}
\usage{
ExtractRegionKeyEndPos(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Extracts the END POS value from a RegionKey.
}
