% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractRegionKeyStartPos}
\alias{ExtractRegionKeyStartPos}
\title{Extracts the START POS value from a RegionKey.}
\usage{
ExtractRegionKeyStartPos(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Extracts the START POS value from a RegionKey.
}
