% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractRegionKeyStrand}
\alias{ExtractRegionKeyStrand}
\title{Extracts the STRAND code from a RegionKey.}
\usage{
ExtractRegionKeyStrand(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Extracts the STRAND code from a RegionKey.
}
