% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractVariantKeyChrom}
\alias{ExtractVariantKeyChrom}
\title{Extracts the CHROM code from a VariantKey.}
\usage{
ExtractVariantKeyChrom(vk)
}
\arguments{
\item{vk}{VariantKey code.}
}
\description{
Extracts the CHROM code from a VariantKey.
}
