% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractVariantKeyPos}
\alias{ExtractVariantKeyPos}
\title{Extracts the POS value from a VariantKey.}
\usage{
ExtractVariantKeyPos(vk)
}
\arguments{
\item{vk}{VariantKey code.}
}
\description{
Extracts the POS value from a VariantKey.
}
