% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractVariantKeyRefAlt}
\alias{ExtractVariantKeyRefAlt}
\title{Extracts the REF+ALT code from a VariantKey.}
\usage{
ExtractVariantKeyRefAlt(vk)
}
\arguments{
\item{vk}{VariantKey code.}
}
\description{
Extracts the REF+ALT code from a VariantKey.
}
