% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FindAllRvVariantKeyByRsid}
\alias{FindAllRvVariantKeyByRsid}
\title{Returns all the VariantKeys associated with an rsID.
NOTE: the output is limited to maximum 10 results.}
\usage{
FindAllRvVariantKeyByRsid(
  rsid,
  max = 10,
  first = 0,
  last = vk.env$rsvk_$NROWS,
  mc = vk.env$rsvk_$MC
)
}
\arguments{
\item{rsid}{rsID to search.}

\item{max}{max number of results to return.}

\item{first}{First element of the range to search (min value = 0).}

\item{last}{Element (up to but not including) where to end the search (max value = nitems).}

\item{mc}{Memory-mapped columns object as returned by MmapRSVKfile.}
}
\description{
Returns all the VariantKeys associated with an rsID.
NOTE: the output is limited to maximum 10 results.
}
