% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FindAllVrRsidByVariantKey}
\alias{FindAllVrRsidByVariantKey}
\title{Returns all the rsIDs associated with a VariantKey.
NOTE: the output is limited to maximum 10 results.}
\usage{
FindAllVrRsidByVariantKey(
  vk,
  max = 10,
  first = 0,
  last = vk.env$vkrs_$NROWS,
  mc = vk.env$vkrs_$MC
)
}
\arguments{
\item{vk}{VariantKey to search.}

\item{max}{max number of results to return.}

\item{first}{First element of the range to search (min value = 0).}

\item{last}{Element (up to but not including) where to end the search (max value = nitems).}

\item{mc}{Memory-mapped columns object as returned by MmapRSVKfile.}
}
\description{
Returns all the rsIDs associated with a VariantKey.
NOTE: the output is limited to maximum 10 results.
}
