% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FindRefAltByVariantKey}
\alias{FindRefAltByVariantKey}
\title{Looks up the REF and ALT strings of a VariantKey.}
\usage{
FindRefAltByVariantKey(vk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey to search.}

\item{mc}{Memory-mapped columns object as returned by MmapNRVKfile.}
}
\value{
REF+ALT length or 0 if the VariantKey is not found.
}
\description{
Looks up the REF and ALT strings of a VariantKey.
}
