% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FindRvVariantKeyByRsid}
\alias{FindRvVariantKeyByRsid}
\title{Returns the first VariantKey associated with an rsID.}
\usage{
FindRvVariantKeyByRsid(
  rsid,
  first = 0,
  last = vk.env$rsvk_$NROWS,
  mc = vk.env$rsvk_$MC
)
}
\arguments{
\item{rsid}{rsID to search.}

\item{first}{First element of the range to search (min value = 0).}

\item{last}{Element (up to but not including) where to end the search (max value = nitems).}

\item{mc}{Memory-mapped columns object as returned by MmapRSVKfile.}
}
\description{
Returns the first VariantKey associated with an rsID.
}
