% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FindVrRsidByVariantKey}
\alias{FindVrRsidByVariantKey}
\title{Returns the first rsID associated with a VariantKey.}
\usage{
FindVrRsidByVariantKey(
  vk,
  first = 0,
  last = vk.env$vkrs_$NROWS,
  mc = vk.env$vkrs_$MC
)
}
\arguments{
\item{vk}{VariantKey.}

\item{first}{First element of the range to search (min value = 0).}

\item{last}{Element (up to but not including) where to end the search (max value = nitems).}

\item{mc}{Memory-mapped columns object as returned by MmapVKRSfile.}
}
\description{
Returns the first rsID associated with a VariantKey.
}
