% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FlipAllele}
\alias{FlipAllele}
\title{Replaces each nucleotide of an allele with its complement.
The resulting string is always in uppercase.
Support extended nucleotide letters.}
\usage{
FlipAllele(allele)
}
\arguments{
\item{allele}{Allele. String containing a sequence of nucleotide letters.}

\item{size}{Length of the allele string.}
}
\description{
Replaces each nucleotide of an allele with its complement.
The resulting string is always in uppercase.
Support extended nucleotide letters.
}
