% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetNextRvVariantKeyByRsid}
\alias{GetNextRvVariantKeyByRsid}
\title{Returns the next VariantKey associated with an rsID.
Call this in a loop after FindRVVariantKeyByRsid to get all the VariantKeys of the same rsID.}
\usage{
GetNextRvVariantKeyByRsid(
  rsid,
  pos,
  last = vk.env$rsvk_$NROWS,
  mc = vk.env$rsvk_$MC
)
}
\arguments{
\item{rsid}{rsID to search.}

\item{pos}{Current item.}

\item{last}{Element (up to but not including) where to end the search (max value = nitems).}

\item{mc}{Memory-mapped columns object as returned by MmapRSVKfile.}
}
\description{
Returns the next VariantKey associated with an rsID.
Call this in a loop after FindRVVariantKeyByRsid to get all the VariantKeys of the same rsID.
}
