% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetNextVrRsidByVariantKey}
\alias{GetNextVrRsidByVariantKey}
\title{Returns the next rsID associated with a VariantKey.
Call this in a loop after FindVRRsidByVariantKey to get all the rsIDs of the same VariantKey.}
\usage{
GetNextVrRsidByVariantKey(
  vk,
  pos,
  last = vk.env$vkrs_$NROWS,
  mc = vk.env$vkrs_$MC
)
}
\arguments{
\item{vk}{VariantKey to search.}

\item{pos}{Current item.}

\item{last}{Element (up to but not including) where to end the search (max value = nitems).}

\item{mc}{Memory-mapped columns object as returned by MmapRSVKfile.}
}
\description{
Returns the next rsID associated with a VariantKey.
Call this in a loop after FindVRRsidByVariantKey to get all the rsIDs of the same VariantKey.
}
