% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetRefLenByVariantKey}
\alias{GetRefLenByVariantKey}
\title{Retrieve the REF length for the specified VariantKey.
Return REF length or 0 if the VariantKey is not reversible and not found.}
\usage{
GetRefLenByVariantKey(src, last, vk)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the VariantKey to REF+ALT lookup table (vknr.bin).}

\item{last}{Number of variants in the src file -1.}

\item{vk}{VariantKey.}
}
\description{
Retrieve the REF length for the specified VariantKey.
Return REF length or 0 if the VariantKey is not reversible and not found.
}
