% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetRegionKeyChromEndPos}
\alias{GetRegionKeyChromEndPos}
\title{Returns the CHROM and END POS of a RegionKey.}
\usage{
GetRegionKeyChromEndPos(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Returns the CHROM and END POS of a RegionKey.
}
