% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetRegionKeyChromStartPos}
\alias{GetRegionKeyChromStartPos}
\title{Returns the CHROM and START POS section of a RegionKey.}
\usage{
GetRegionKeyChromStartPos(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Returns the CHROM and START POS section of a RegionKey.
}
