% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetRvVariantkey}
\alias{GetRvVariantkey}
\title{Returns the VariantKey at the specified position in the RV file.}
\usage{
GetRvVariantkey(src, item)
}
\arguments{
\item{src}{Address of the memory mapped binary file containing the rsID to VariantKey lookup table (rsvk.bin).}

\item{item}{Binary block number.}
}
\description{
Returns the VariantKey at the specified position in the RV file.
}
