% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVariantKeyChromEndPos}
\alias{GetVariantKeyChromEndPos}
\title{Returns the CHROM and END POS of a VariantKey.}
\usage{
GetVariantKeyChromEndPos(vk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey.}

\item{mc}{Memory-mapped columns object as returned by MmapNRVKfile.}
}
\description{
Returns the CHROM and END POS of a VariantKey.
}
