% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVariantKeyChromStartPos}
\alias{GetVariantKeyChromStartPos}
\title{Returns the CHROM and START POS section of a VariantKey.}
\usage{
GetVariantKeyChromStartPos(vk)
}
\arguments{
\item{vk}{VariantKey.}
}
\description{
Returns the CHROM and START POS section of a VariantKey.
}
