% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVariantKeyEndPos}
\alias{GetVariantKeyEndPos}
\title{Returns the end position of a VariantKey (POS + REF length).}
\usage{
GetVariantKeyEndPos(vk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey.}

\item{mc}{Memory-mapped columns object as returned by MmapNRVKfile.}
}
\value{
Variant end position.
}
\description{
Returns the end position of a VariantKey (POS + REF length).
}
