% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVariantKeyRefLength}
\alias{GetVariantKeyRefLength}
\title{Returns the REF length of a VariantKey.}
\usage{
GetVariantKeyRefLength(vk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey.}

\item{mc}{Memory-mapped columns object as returned by MmapNRVKfile.}
}
\value{
REF length or 0 if the VariantKey is not reversible and not found.
}
\description{
Returns the REF length of a VariantKey.
}
