% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVrRsid}
\alias{GetVrRsid}
\title{Returns the RSID at the specified position in the VR file.}
\usage{
GetVrRsid()
}
\arguments{
\item{src}{Address of the memory mapped binary file containing the VariantKey to rsID lookup table (vkrs.bin).}

\item{item}{Binary block number.}
}
\description{
Returns the RSID at the specified position in the VR file.
}
