% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{>=.uint64}
\alias{>=.uint64}
\title{Returns true if x is greater or equal than y.}
\usage{
\method{>=}{uint64}(x, y)
}
\arguments{
\item{x}{uint64 vector}

\item{y}{uint64 vector}
}
\description{
Returns true if x is greater or equal than y.
}
