% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{HashStringID}
\alias{HashStringID}
\title{Hashes a string into a non-reversible 64 bit string ID.
This function can be used to convert long string IDs to numeric IDs.}
\usage{
HashStringID(str)
}
\arguments{
\item{str}{The string to encode.}
}
\description{
Hashes a string into a non-reversible 64 bit string ID.
This function can be used to convert long string IDs to numeric IDs.
}
