% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex64.R
\name{identical.hex64}
\alias{identical.hex64}
\title{Identity function for class hex64.}
\usage{
identical.hex64(x, y, num.eq = FALSE, single.NA = FALSE,
  attrib.as.set = TRUE, ignore.bytecode = TRUE)
}
\description{
Identity function for class hex64.
}
