% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{intersect.uint64}
\alias{intersect.uint64}
\title{Returns the intersection of two sorted uint64 vectors.}
\usage{
intersect.uint64(x, y)
}
\arguments{
\item{x}{uint64 vector}

\item{y}{uint64 vector}
}
\description{
Returns the intersection of two sorted uint64 vectors.
}
