% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{is.na.uint64}
\alias{is.na.uint64}
\title{Missing values of a uint64 vector.}
\usage{
\method{is.na}{uint64}(x)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
A uint64 vector has no missing values, so FALSE is always returned. A uint64
is stored in the bit pattern of a double, which the default is.na() would
report as missing when all the exponent bits are set.
}
