% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{length<-.uint64}
\alias{length<-.uint64}
\title{Set the length of uint64 vector.}
\usage{
\method{length}{uint64}(x) <- value
}
\arguments{
\item{x}{uint64 vector}

\item{value}{value to set the new length}
}
\description{
Set the length of uint64 vector.
}
