% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{LoadGenorefIndex}
\alias{LoadGenorefIndex}
\title{Load the genome reference index.}
\usage{
LoadGenorefIndex(src)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the genome reference data (fasta.bin).}

\item{idx}{Index of sequences offset by chromosome number (1 to 25). The index 26 contains the file length. Requires 27 elements.}
}
\description{
Load the genome reference index.
}
