% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{MmapBinfile}
\alias{MmapBinfile}
\title{Memory map the specified file.}
\usage{
MmapBinfile(file)
}
\arguments{
\item{file}{Path to the file to map.}

\item{mf}{Structure containing the memory mapped file.}
}
\description{
Memory map the specified file.
}
