% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{MmapGenorefFile}
\alias{MmapGenorefFile}
\title{Memory maps the genoref binary file (fasta.bin).}
\usage{
MmapGenorefFile(file)
}
\arguments{
\item{file}{Path to the file to map.}
}
\value{
The memory mapped file object, the columns object and the number of rows.
}
\description{
Memory maps the genoref binary file (fasta.bin).
}
