% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{MmapNRVKFile}
\alias{MmapNRVKFile}
\title{Memory maps the NRVK binary file (nrvk.bin).}
\usage{
MmapNRVKFile(file)
}
\arguments{
\item{file}{Path to the file to map.}
}
\value{
The memory mapped file object, the columns object and the number of rows.
}
\description{
Memory maps the NRVK binary file (nrvk.bin).
}
