% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{MmapVKRSFile}
\alias{MmapVKRSFile}
\title{Memory maps the VKRS binary file (vkrs.bin).}
\usage{
MmapVKRSFile(file, ctbytes)
}
\arguments{
\item{file}{Path to the file to map.}

\item{ctbytes}{List containing the number of bytes for each column type (i.e. 1 for uint8, 2 for uint16, 4 for uint32, 8 for uint64)}
}
\value{
The memory mapped file object, the columns object and the number of rows.
}
\description{
Memory maps the VKRS binary file (vkrs.bin).
}
