% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{MunmapBinfile}
\alias{MunmapBinfile}
\title{Unmaps and closes a memory mapped file.
On success, munmap() returns 0, on failure -1.}
\usage{
MunmapBinfile(mf)
}
\arguments{
\item{mf}{Descriptor of memory-mapped file.}
}
\description{
Unmaps and closes a memory mapped file.
On success, munmap() returns 0, on failure -1.
}
