% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{NormalizedVariantKey}
\alias{NormalizedVariantKey}
\title{Normalizes a variant and returns its VariantKey.}
\usage{
NormalizedVariantKey(chrom, pos, posindex, ref, alt, mf = vk.env$genoref_$MF)
}
\arguments{
\item{chrom}{Chromosome encoded number.}

\item{pos}{Position. The reference position.}

\item{posindex}{Position index: 0 for 0-based, 1 for 1-based.}

\item{ref}{Reference allele. String containing a sequence of nucleotide letters.}

\item{alt}{Alternate non-reference allele string.}

\item{mf}{Memory-mapped file object as returned by MmapGenorefFile.}
}
\description{
Normalizes a variant and returns its VariantKey.
}
