% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{NormalizeVariant}
\alias{NormalizeVariant}
\title{Normalizes a variant against the genome reference, flipping the alleles if required.
See the normalization algorithm described at:
https://genome.sph.umich.edu/wiki/Variant_Normalization}
\usage{
NormalizeVariant(chrom, pos, ref, alt, mf = vk.env$genoref_$MF)
}
\arguments{
\item{chrom}{Chromosome encoded number.}

\item{pos}{Position. The reference position, with the first base having position 0.}

\item{ref}{Reference allele. String containing a sequence of nucleotide letters.}

\item{alt}{Alternate non-reference allele string.}

\item{mf}{Memory-mapped file object as returned by MmapGenorefFile.}
}
\value{
Positive bitmask number in case of success, negative number in case of error.
  When positive, each bit set has a different meaning, as defined by the NORM_* codes:
  * bit 0:
  Reference allele is inconsistent with the genome reference (i.e. when contains nucleotide letters other than A, C, G and T).
  * bit 1:
  Alleles have been swapped.
  * bit 2:
  Allele nucleotides have been flipped (each nucleotide has been replaced with its complement).
  * bit 3:
  Alleles have been left extended.
  * bit 4:
  Alleles have been right trimmed.
  * bit 5:
  Alleles have been left trimmed.
}
\description{
Normalizes a variant against the genome reference, flipping the alleles if required.
See the normalization algorithm described at:
https://genome.sph.umich.edu/wiki/Variant_Normalization
}
