% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{order.uint64}
\alias{order.uint64}
\title{Returns a permutation which rearranges its first argument into ascending order.}
\usage{
order.uint64(x)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Returns a permutation which rearranges its first argument into ascending order.
}
