% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ParseRegionKeyHex}
\alias{ParseRegionKeyHex}
\title{Parses a 16 character hexadecimal string into a RegionKey.}
\usage{
ParseRegionKeyHex(hex)
}
\arguments{
\item{hex}{RegionKey hexadecimal string (it must contain 16 hexadecimal characters).}
}
\description{
Parses a 16 character hexadecimal string into a RegionKey.
}
