% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ParseVariantKeyHex}
\alias{ParseVariantKeyHex}
\title{Parses a 16 character hexadecimal string into a VariantKey.}
\usage{
ParseVariantKeyHex(hex)
}
\arguments{
\item{hex}{VariantKey hexadecimal string (it must contain 16 hexadecimal characters).}
}
\description{
Parses a 16 character hexadecimal string into a VariantKey.
}
