% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{RegionKeyHex}
\alias{RegionKeyHex}
\title{Returns a RegionKey as a 16 character hexadecimal string.}
\usage{
RegionKeyHex(vk)
}
\arguments{
\item{vk}{RegiontKey code.}
}
\description{
Returns a RegionKey as a 16 character hexadecimal string.
}
