% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{rep.uint64}
\alias{rep.uint64}
\title{Replicate elements of uint64 vectors.}
\usage{
\method{rep}{uint64}(x, ...)
}
\arguments{
\item{x}{uint64 vector to be replicated}
}
\description{
Replicate elements of uint64 vectors.
}
