% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ReverseRegionKey}
\alias{ReverseRegionKey}
\title{Reverses a RegionKey into its decoded components.}
\usage{
ReverseRegionKey(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Reverses a RegionKey into its decoded components.
}
