% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ReverseVariantKey}
\alias{ReverseVariantKey}
\title{Reverses a VariantKey into its CHROM, POS, REF and ALT components.}
\usage{
ReverseVariantKey(vk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey code.}

\item{mc}{Memory-mapped columns object as returned by MmapNRVKfile.}
}
\description{
Reverses a VariantKey into its CHROM, POS, REF and ALT components.
}
