% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{setdiff.uint64}
\alias{setdiff.uint64}
\title{Performs set setdiff on two vectors.}
\usage{
setdiff.uint64(x, y)
}
\arguments{
\item{x}{uint64 vector}

\item{y}{uint64 vector}
}
\description{
Performs set setdiff on two vectors.
}
