% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{setequal.uint64}
\alias{setequal.uint64}
\title{Performs set equality on two vectors.}
\usage{
setequal.uint64(x, y)
}
\arguments{
\item{x}{uint64 vector}

\item{y}{uint64 vector}
}
\description{
Performs set equality on two vectors.
}
