% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{union.uint64}
\alias{union.uint64}
\title{Performs set union on two vectors.}
\usage{
union.uint64(x, y)
}
\arguments{
\item{x}{uint64 vector}

\item{y}{uint64 vector}
}
\description{
Performs set union on two vectors.
}
