% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VariantKey}
\alias{VariantKey}
\title{Returns a VariantKey for the given CHROM, POS (0-based), REF and ALT.
The variant should be already normalized (see NormalizeVariant or use NormalizedVariantkey).}
\usage{
VariantKey(chrom, pos, ref, alt)
}
\arguments{
\item{chrom}{Chromosome. An identifier from the reference genome, no white-space or leading zeros permitted.}

\item{pos}{Position. The reference position, with the first base having position 0.}

\item{ref}{Reference allele. String containing a sequence of nucleotide letters.}

\item{alt}{Alternate non-reference allele string.}
}
\description{
Returns a VariantKey for the given CHROM, POS (0-based), REF and ALT.
The variant should be already normalized (see NormalizeVariant or use NormalizedVariantkey).
}
