% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VariantKeyHex}
\alias{VariantKeyHex}
\title{Returns a VariantKey as a 16 character hexadecimal string.}
\usage{
VariantKeyHex(vk)
}
\arguments{
\item{vk}{VariantKey code.}
}
\description{
Returns a VariantKey as a 16 character hexadecimal string.
}
