% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VariantKeyRange}
\alias{VariantKeyRange}
\title{Returns the minimum and maximum VariantKey of a CHROM and POS range.}
\usage{
VariantKeyRange(chrom, pos_min, pos_max)
}
\arguments{
\item{chrom}{Chromosome numerical code.}

\item{pos_min}{Start reference position, with the first base having position 0.}

\item{pos_max}{End reference position, with the first base having position 0.}
}
\description{
Returns the minimum and maximum VariantKey of a CHROM and POS range.
}
