% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VariantToRegionkey}
\alias{VariantToRegionkey}
\title{Converts a VariantKey into a RegionKey.}
\usage{
VariantToRegionkey(vk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey code.}

\item{mc}{Memory-mapped columns object as returned by MmapNRVKfile.}
}
\description{
Converts a VariantKey into a RegionKey.
}
