% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{xtfrm.uint64}
\alias{xtfrm.uint64}
\title{Comparison key for a uint64 vector.}
\usage{
\method{xtfrm}{uint64}(x)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Provides the ordering used by order(), sort.list() and rank(), which would
otherwise compare the underlying doubles, i.e. the raw 64 bit patterns.
}
