% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{AreOverlappingRegionKeys}
\alias{AreOverlappingRegionKeys}
\title{Check if two regionkeys are overlapping.
Return 1 if the regions overlap, 0 otherwise.}
\usage{
AreOverlappingRegionKeys(rka, rkb)
}
\arguments{
\item{rka}{RegionKey A.}

\item{rkb}{RegionKey B.}
}
\description{
Check if two regionkeys are overlapping.
Return 1 if the regions overlap, 0 otherwise.
}
