% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{AreOverlappingVariantKeyRegionKey}
\alias{AreOverlappingVariantKeyRegionKey}
\title{Check if variantkey and regionkey are overlapping.
Return 1 if the regions overlap, 0 otherwise.}
\usage{
AreOverlappingVariantKeyRegionKey(src, last, vk, rk)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the VariantKey to REF+ALT lookup table (vknr.bin).}

\item{last}{Number of variants in the src file -1.}

\item{vk}{VariantKey code.}

\item{rk}{RegionKey code.}
}
\description{
Check if variantkey and regionkey are overlapping.
Return 1 if the regions overlap, 0 otherwise.
}
