% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{AreOverlappingVariantKeyRegionKey}
\alias{AreOverlappingVariantKeyRegionKey}
\title{Check if variantkey and regionkey are overlapping.
Return 1 if the regions overlap, 0 otherwise.}
\usage{
AreOverlappingVariantKeyRegionKey(vk, rk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey code.}

\item{rk}{RegionKey code.}

\item{mc}{Memory-mapped columns object as retured by MmapNRVKfile.}
}
\description{
Check if variantkey and regionkey are overlapping.
Return 1 if the regions overlap, 0 otherwise.
}
