% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.hex64.uint64}
\alias{as.hex64.uint64}
\title{Coerce hex64 vector vector to uint64.}
\usage{
\method{as.hex64}{uint64}(x, ...)
}
\arguments{
\item{x}{hexadecimal character vector (16 characters per item)}
}
\description{
Coerce hex64 vector vector to uint64.
}
