% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{as.uint64.factor}
\alias{as.uint64.factor}
\title{Coerce from factor to uint64.}
\usage{
\method{as.uint64}{factor}(x, ...)
}
\arguments{
\item{x}{factor vector}
}
\description{
Coerce from factor to uint64.
}
