% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{CheckReference}
\alias{CheckReference}
\title{Check if the reference allele matches the reference genome data.
Return Positive number in case of success, negative in case of error:
  *  0 the reference allele match the reference genome;
  *  1 the reference allele is inconsistent with the genome reference (i.e. when contains nucleotide letters other than A, C, G and T);
  * -1 the reference allele don't match the reference genome;
  * -2 the reference allele is longer than the genome reference sequence.}
\usage{
CheckReference()
}
\arguments{
\item{src}{Address of the memory mapped input file containing the genome reference data (fasta.bin).}

\item{idx}{Index of sequences offset by chromosome number (1 to 25).}

\item{chrom}{Encoded Chromosome number (see encode_chrom).}

\item{pos}{Position. The reference position, with the first base having position 0.}

\item{ref}{Reference allele. String containing a sequence of nucleotide letters.}
}
\description{
Check if the reference allele matches the reference genome data.
Return Positive number in case of success, negative in case of error:
  *  0 the reference allele match the reference genome;
  *  1 the reference allele is inconsistent with the genome reference (i.e. when contains nucleotide letters other than A, C, G and T);
  * -1 the reference allele don't match the reference genome;
  * -2 the reference allele is longer than the genome reference sequence.
}
