% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{CloseVariantKey}
\alias{CloseVariantKey}
\title{Unmap the memory-mapped files.
This should be the last function called in order to close any open memory-mapped file.}
\usage{
CloseVariantKey()
}
\description{
Unmap the memory-mapped files.
This should be the last function called in order to close any open memory-mapped file.
}
