% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{CompareVariantKeyChrom}
\alias{CompareVariantKeyChrom}
\title{Compares two VariantKeys by chromosome only.
Returns -1 if the first chromosome is smaller than the second, 0 if they are equal and 1 if the first is greater than the second.}
\usage{
CompareVariantKeyChrom(vka, vkb)
}
\arguments{
\item{vka}{The first VariantKey to be compared.}

\item{vkb}{The second VariantKey to be compared.}
}
\description{
Compares two VariantKeys by chromosome only.
Returns -1 if the first chromosome is smaller than the second, 0 if they are equal and 1 if the first is greater than the second.
}
