% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{DecodeRefAlt}
\alias{DecodeRefAlt}
\title{Decode the 32 bit REF+ALT code if reversible (if it has 11 or less bases in total and only contains ACGT letters).}
\usage{
DecodeRefAlt(code)
}
\arguments{
\item{code}{REF+ALT code}
}
\description{
Decode the 32 bit REF+ALT code if reversible (if it has 11 or less bases in total and only contains ACGT letters).
}
