% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{DecodeRegionKey}
\alias{DecodeRegionKey}
\title{Decode a RegionKey code and returns the components as regionkey_t structure.}
\usage{
DecodeRegionKey(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Decode a RegionKey code and returns the components as regionkey_t structure.
}
