% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{DecodeRegionStrand}
\alias{DecodeRegionStrand}
\title{Decode the strand direction code (0 > 0, 1 > +1, 2 > -1).}
\usage{
DecodeRegionStrand(strand)
}
\arguments{
\item{code}{Strand code.}
}
\description{
Decode the strand direction code (0 > 0, 1 > +1, 2 > -1).
}
