% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{DecodeStringID}
\alias{DecodeStringID}
\title{Decode the encoded string ID.
This function is the reverse of encode_string_id.
The string is always returned in uppercase mode.}
\usage{
DecodeStringID(esid)
}
\arguments{
\item{esid}{Encoded string ID code.}
}
\description{
Decode the encoded string ID.
This function is the reverse of encode_string_id.
The string is always returned in uppercase mode.
}
