% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{DecodeVariantkey}
\alias{DecodeVariantkey}
\title{Decode a VariantKey code and returns the components.}
\usage{
DecodeVariantkey(vk)
}
\arguments{
\item{vk}{Decoded variantkey structure.}

\item{code}{VariantKey code.}
}
\description{
Decode a VariantKey code and returns the components.
}
