% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{EncodeRefAlt}
\alias{EncodeRefAlt}
\title{Returns reference+alternate encoding.}
\usage{
EncodeRefAlt(ref, alt)
}
\arguments{
\item{ref}{Reference allele. String containing a sequence of nucleotide letters.}

\item{alt}{Alternate non-reference allele string.}
}
\description{
Returns reference+alternate encoding.
}
