% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{EncodeRegionKey}
\alias{EncodeRegionKey}
\title{Returns a 64 bit regionkey}
\usage{
EncodeRegionKey(chrom, startpos, endpos, strand)
}
\arguments{
\item{chrom}{Encoded Chromosome (see encode_chrom).}

\item{startpos}{Start position (zero based).}

\item{endpos}{End position (startpos + region_length).}

\item{strand}{Encoded Strand direction (-1 > 2, 0 > 0, +1 > 1)}
}
\description{
Returns a 64 bit regionkey
}
