% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{EncodeRegionStrand}
\alias{EncodeRegionStrand}
\title{Encode the strand direction (-1 > 2, 0 > 0, +1 > 1).}
\usage{
EncodeRegionStrand(strand)
}
\arguments{
\item{strand}{Strand direction (-1, 0, +1).}
}
\description{
Encode the strand direction (-1 > 2, 0 > 0, +1 > 1).
}
