% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{EncodeStringID}
\alias{EncodeStringID}
\title{Encode maximum 10 characters of a string into a 64 bit unsigned integer.
This function can be used to convert generic string IDs to numeric IDs.}
\usage{
EncodeStringID(str, start = 0)
}
\arguments{
\item{str}{The string to encode. The characters beyond the first 10 from start are ignored. It supports ASCII characters from '!' to 'z'.}

\item{start}{First character to encode, starting from 0. To encode the last 10 characters, set this value at (size - 10).}
}
\description{
Encode maximum 10 characters of a string into a 64 bit unsigned integer.
This function can be used to convert generic string IDs to numeric IDs.
}
