% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{EncodeStringNumID}
\alias{EncodeStringNumID}
\title{Encode a string composed by a character section followed by a separator character and a numerical section
into a 64 bit unsigned integer. For example: "ABCDE:0001234"
Encodes up to 5 characters in uppercase, a number up to 2^27, and up to 7 zero padding digits.
If the string is 10 character or less, then the EncodeStringID() is used.}
\usage{
EncodeStringNumID(str, sep = ":")
}
\arguments{
\item{str}{The string to encode. It supports ASCII characters from '!' to 'z'.}

\item{sep}{Separator character between string and number.}
}
\description{
Encode a string composed by a character section followed by a separator character and a numerical section
into a 64 bit unsigned integer. For example: "ABCDE:0001234"
Encodes up to 5 characters in uppercase, a number up to 2^27, and up to 7 zero padding digits.
If the string is 10 character or less, then the EncodeStringID() is used.
}
