% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{EncodeVariantKey}
\alias{EncodeVariantKey}
\title{Returns a 64 bit variant key based on pre-encoded CHROM, POS (0-base) and REF+ALT.}
\usage{
EncodeVariantKey(chrom, pos, refalt)
}
\arguments{
\item{chrom}{Encoded Chromosome (see EncodeChrom)}

\item{pos}{Position. The reference position, with the first base having position 0.}

\item{refalt}{Encoded Reference + Alternate (see EncodeRefAlt)}
}
\description{
Returns a 64 bit variant key based on pre-encoded CHROM, POS (0-base) and REF+ALT.
}
