% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtendRegionKey}
\alias{ExtendRegionKey}
\title{Extend a regionkey region by a fixed amount from the start and end position.}
\usage{
ExtendRegionKey(rk, size)
}
\arguments{
\item{rk}{RegionKey code.}

\item{size}{Amount to extend the region.}
}
\description{
Extend a regionkey region by a fixed amount from the start and end position.
}
