% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractRegionKeyChrom}
\alias{ExtractRegionKeyChrom}
\title{Extract the CHROM code from RegionKey.}
\usage{
ExtractRegionKeyChrom(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Extract the CHROM code from RegionKey.
}
