% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractRegionKeyEndPos}
\alias{ExtractRegionKeyEndPos}
\title{Extract the END POS code from RegionKey.}
\usage{
ExtractRegionKeyEndPos(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Extract the END POS code from RegionKey.
}
