% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractRegionKeyStartPos}
\alias{ExtractRegionKeyStartPos}
\title{Extract the START POS code from RegionKey.}
\usage{
ExtractRegionKeyStartPos(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Extract the START POS code from RegionKey.
}
