% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractRegionKeyStrand}
\alias{ExtractRegionKeyStrand}
\title{Extract the STRAND from RegionKey.}
\usage{
ExtractRegionKeyStrand(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Extract the STRAND from RegionKey.
}
