% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractVariantKeyChrom}
\alias{ExtractVariantKeyChrom}
\title{Extract the CHROM code from VariantKey.}
\usage{
ExtractVariantKeyChrom(vk)
}
\arguments{
\item{vk}{VariantKey code.}
}
\description{
Extract the CHROM code from VariantKey.
}
