% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractVariantkeyChrom}
\alias{ExtractVariantkeyChrom}
\title{Extract the CHROM code from VariantKey.}
\usage{
ExtractVariantkeyChrom(vk)
}
\arguments{
\item{vk}{VariantKey code.}
}
\description{
Extract the CHROM code from VariantKey.
}
