% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractVariantKeyPos}
\alias{ExtractVariantKeyPos}
\title{Extract the POS code from VariantKey.}
\usage{
ExtractVariantKeyPos(vk)
}
\arguments{
\item{vk}{VariantKey code.}
}
\description{
Extract the POS code from VariantKey.
}
