% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractVariantkeyPos}
\alias{ExtractVariantkeyPos}
\title{Extract the POS code from VariantKey.}
\usage{
ExtractVariantkeyPos(vk)
}
\arguments{
\item{vk}{VariantKey code.}
}
\description{
Extract the POS code from VariantKey.
}
