% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ExtractVariantkeyRefAlt}
\alias{ExtractVariantkeyRefAlt}
\title{Extract the REF+ALT code from VariantKey.}
\usage{
ExtractVariantkeyRefAlt(vk)
}
\arguments{
\item{vk}{VariantKey code.}
}
\description{
Extract the REF+ALT code from VariantKey.
}
