% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FindRefAltByVariantkey}
\alias{FindRefAltByVariantkey}
\title{Retrieve the REF and ALT strings for the specified VariantKey.
Return REF+ALT length or 0 if the VariantKey is not found.}
\usage{
FindRefAltByVariantkey(src, last, vk)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the VariantKey to REF+ALT lookup table (vknr.bin).}

\item{last}{Number of variants in the src file -1.}

\item{vk}{VariantKey to search.}
}
\description{
Retrieve the REF and ALT strings for the specified VariantKey.
Return REF+ALT length or 0 if the VariantKey is not found.
}
