% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FindRefAltByVariantKey}
\alias{FindRefAltByVariantKey}
\title{Retrieve the REF and ALT strings for the specified VariantKey.
Return REF+ALT length or 0 if the VariantKey is not found.}
\usage{
FindRefAltByVariantKey(vk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey to search.}

\item{mc}{Memory-mapped columns object as retured by MmapNRVKfile.}
}
\description{
Retrieve the REF and ALT strings for the specified VariantKey.
Return REF+ALT length or 0 if the VariantKey is not found.
}
