% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FindRvVariantKeyByRsid}
\alias{FindRvVariantKeyByRsid}
\title{Search for the specified rsID and returns the first occurrence of VariantKey in the RV file, or zero if not found.}
\usage{
FindRvVariantKeyByRsid(mc, first, last, rsid)
}
\arguments{
\item{mc}{Memory-mapped columns object as retured by MmapRSVKfile.}

\item{first}{First element of the range to search (min value = 0).}

\item{last}{Element (up to but not including) where to end the search (max value = nitems).}

\item{rsid}{rsID to search.}
}
\description{
Search for the specified rsID and returns the first occurrence of VariantKey in the RV file, or zero if not found.
}
