% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FindVrChromposRange}
\alias{FindVrChromposRange}
\title{Search for the specified CHROM-POS range and returns the first occurrence of rsID in the VR file.}
\usage{
FindVrChromposRange(src, first, last, chrom, pos_min, pos_max)
}
\arguments{
\item{src}{Address of the memory mapped binary file containing the VariantKey to rsID lookup table (vkrs.bin).}

\item{first}{First element of the range to search (min value = 0).}

\item{last}{Last element of the range to search (max value = nitems - 1).}

\item{chrom}{Chromosome encoded number.}

\item{pos_min}{Start reference position, with the first base having position 0.}

\item{pos_max}{End reference position, with the first base having position 0.}
}
\description{
Search for the specified CHROM-POS range and returns the first occurrence of rsID in the VR file.
}
