% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{FindVrRsidByVariantKey}
\alias{FindVrRsidByVariantKey}
\title{Search for the specified VariantKey and returns the first occurrence of rsID in the VR file, or 0 if not found}
\usage{
FindVrRsidByVariantKey(mc, first, last, vk)
}
\arguments{
\item{mc}{Memory-mapped columns object as retured by MmapVKRSfile.}

\item{first}{First element of the range to search (min value = 0).}

\item{last}{Last element of the range to search (max value = nitems - 1).}

\item{vk}{VariantKey.}
}
\description{
Search for the specified VariantKey and returns the first occurrence of rsID in the VR file, or 0 if not found
}
