% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetGenorefSeq}
\alias{GetGenorefSeq}
\title{Returns the genome reference nucleotide at the specified chromosome and position, or 0 in case of invalid position.}
\usage{
GetGenorefSeq(mf, chrom, pos)
}
\arguments{
\item{chrom}{Encoded Chromosome number (see encode_chrom).}

\item{pos}{Position. The reference position, with the first base having position 0.}

\item{mc}{Memory-mapped columns object as retured by MmapGenorefFile.}
}
\description{
Returns the genome reference nucleotide at the specified chromosome and position, or 0 in case of invalid position.
}
