% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetGenorefSeq}
\alias{GetGenorefSeq}
\title{Returns the genome reference nucleotide at the specified chromosome and position.}
\usage{
GetGenorefSeq(chrom, pos, mf = vk.env$genoref_$MF)
}
\arguments{
\item{chrom}{Encoded Chromosome number (see encode_chrom).}

\item{pos}{Position. The reference position, with the first base having position 0.}

\item{mf}{Memory-mapped file object as retured by MmapGenorefFile.}
}
\description{
Returns the genome reference nucleotide at the specified chromosome and position.
}
