% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetGenorefSeq}
\alias{GetGenorefSeq}
\title{Returns the genome reference nucleotide at the specified chromosome and position, or 0 in case of invalid position.}
\usage{
GetGenorefSeq(src, idx, chrom, pos)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the genome reference data (fasta.bin).}

\item{idx}{Index of sequences offset by chromosome number (1 to 25).}

\item{chrom}{Encoded Chromosome number (see encode_chrom).}

\item{pos}{Position. The reference position, with the first base having position 0.}
}
\description{
Returns the genome reference nucleotide at the specified chromosome and position, or 0 in case of invalid position.
}
