% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetNextRvVariantKeyByRsid}
\alias{GetNextRvVariantKeyByRsid}
\title{Get the next VariantKey for the specified rsID in the RV file, or 0 if not found
This function can be called in a loop to get all VariantKeys that are associated with the same rsID (if any).}
\usage{
GetNextRvVariantKeyByRsid(mc, pos, last, rsid)
}
\arguments{
\item{mc}{Memory-mapped columns object as retured by MmapRSVKfile.}

\item{pos}{Current item.}

\item{last}{Element (up to but not including) where to end the search (max value = nitems).}

\item{rsid}{rsID to search.}
}
\description{
Get the next VariantKey for the specified rsID in the RV file, or 0 if not found
This function can be called in a loop to get all VariantKeys that are associated with the same rsID (if any).
}
