% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetNextRvVariantkeyByRsid}
\alias{GetNextRvVariantkeyByRsid}
\title{Get the next VariantKey for the specified rsID in the RV file, or 0 if not found
This function can be called in a loop to get all VariantKeys that are associated with the same rsID (if any).}
\usage{
GetNextRvVariantkeyByRsid()
}
\arguments{
\item{src}{Address of the memory mapped binary file containing the rsID to VariantKey lookup table (rsvk.bin).}

\item{pos}{Current item.}

\item{last}{Last element of the range to search (max value = nitems - 1).}

\item{rsid}{rsID to search.}
}
\description{
Get the next VariantKey for the specified rsID in the RV file, or 0 if not found
This function can be called in a loop to get all VariantKeys that are associated with the same rsID (if any).
}
