% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetNextVrRsidByVariantKey}
\alias{GetNextVrRsidByVariantKey}
\title{Get the next rsID for the specified VariantKey in the VR file, or 0 if not found
This function can be called in a loop to get all rsIDs that are associated with the same VariantKey (if any).}
\usage{
GetNextVrRsidByVariantKey(vk, pos, last = vk.env$vkrs_$NROWS,
  mc = vk.env$vkrs_$MC)
}
\arguments{
\item{vk}{VariantKey to search.}

\item{pos}{Current item.}

\item{last}{Element (up to but not including) where to end the search (max value = nitems).}

\item{mc}{Memory-mapped columns object as retured by MmapRSVKfile.}
}
\description{
Get the next rsID for the specified VariantKey in the VR file, or 0 if not found
This function can be called in a loop to get all rsIDs that are associated with the same VariantKey (if any).
}
