% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetRegionKeyChromEndPos}
\alias{GetRegionKeyChromEndPos}
\title{Get the CHROM + END POS encoding from RegionKey.}
\usage{
GetRegionKeyChromEndPos(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Get the CHROM + END POS encoding from RegionKey.
}
