% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetRegionKeyChromStartPos}
\alias{GetRegionKeyChromStartPos}
\title{Get the CHROM + START POS encoding from RegionKey.}
\usage{
GetRegionKeyChromStartPos(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Get the CHROM + START POS encoding from RegionKey.
}
