% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVariantKeyChromEndPos}
\alias{GetVariantKeyChromEndPos}
\title{Get the CHROM + END POS encoding from VariantKey.}
\usage{
GetVariantKeyChromEndPos(src, last, vk)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the VariantKey to REF+ALT lookup table (vknr.bin).}

\item{last}{Number of variants in the src file -1.}

\item{vk}{VariantKey.}
}
\description{
Get the CHROM + END POS encoding from VariantKey.
}
