% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVariantKeyChromEndPos}
\alias{GetVariantKeyChromEndPos}
\title{Get the CHROM + END POS encoding from VariantKey.}
\usage{
GetVariantKeyChromEndPos(vk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey.}

\item{mc}{Memory-mapped columns object as retured by MmapNRVKfile.}
}
\description{
Get the CHROM + END POS encoding from VariantKey.
}
