% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVariantKeyChromStartPos}
\alias{GetVariantKeyChromStartPos}
\title{Get the CHROM + START POS encoding from VariantKey.}
\usage{
GetVariantKeyChromStartPos(vk)
}
\arguments{
\item{vk}{VariantKey.}
}
\description{
Get the CHROM + START POS encoding from VariantKey.
}
