% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVariantKeyEndPos}
\alias{GetVariantKeyEndPos}
\title{Get the VariantKey end position (POS + REF length).
Return variant end position.}
\usage{
GetVariantKeyEndPos(mc, vk)
}
\arguments{
\item{mc}{Memory-mapped columns object as retured by MmapNRVKfile.}

\item{vk}{VariantKey.}
}
\description{
Get the VariantKey end position (POS + REF length).
Return variant end position.
}
