% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVariantKeyEndPos}
\alias{GetVariantKeyEndPos}
\title{Get the VariantKey end position (POS + REF length).
Return variant end position.}
\usage{
GetVariantKeyEndPos(src, last, vk)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the VariantKey to REF+ALT lookup table (nrvk.bin).}

\item{last}{Number of variants in the src file -1.}

\item{vk}{VariantKey.}
}
\description{
Get the VariantKey end position (POS + REF length).
Return variant end position.
}
