% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{GetVariantKeyRefLength}
\alias{GetVariantKeyRefLength}
\title{Retrieve the REF length for the specified VariantKey.
Return REF length or 0 if the VariantKey is not reversible and not found.}
\usage{
GetVariantKeyRefLength(mc, vk)
}
\arguments{
\item{mc}{Memory-mapped columns object as retured by MmapNRVKfile.}

\item{vk}{VariantKey.}
}
\description{
Retrieve the REF length for the specified VariantKey.
Return REF length or 0 if the VariantKey is not reversible and not found.
}
