% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{HashStringID}
\alias{HashStringID}
\title{Hash the input string into a 64 bit unsigned integer.
This function can be used to convert long string IDs to numeric IDs.}
\usage{
HashStringID(str)
}
\arguments{
\item{str}{The string to encode.}
}
\description{
Hash the input string into a 64 bit unsigned integer.
This function can be used to convert long string IDs to numeric IDs.
}
