% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{hexToUint64}
\alias{hexToUint64}
\title{Convert hexadecimal character vector to uint64}
\usage{
hexToUint64(x, ...)
}
\arguments{
\item{x}{hexadecimal character vector (16 characters per item)}
}
\description{
Convert hexadecimal character vector to uint64
}
