% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{identical.uint64}
\alias{identical.uint64}
\title{Identity function for class uint64.}
\usage{
identical.uint64(x, y, num.eq = FALSE, single.NA = FALSE,
  attrib.as.set = TRUE, ignore.bytecode = TRUE)
}
\description{
Identity function for class uint64.
}
