% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{InitVariantKey}
\alias{InitVariantKey}
\title{Load the VariantKey support files.
This should be the first function called in order to load the support files.}
\usage{
InitVariantKey(genoref_file = "", nrvk_file = "", rsvk_file = "",
  vkrs_file = "")
}
\arguments{
\item{genoref_file}{Name and path of the binary file containing the genome reference (fasta.bin). This file can be generated from a FASTA file using the resources/tools/fastabin.sh script.}

\item{nrvk_file}{Name and path of the binary file containing the non-reversible-VariantKey mapping (nrvk.bin). This file can be generated from a normalized VCF file using the resources/tools/nrvk.sh script.}

\item{rsvk_file}{Name and path of the binary file containing the rsID to VariantKey mapping (rsvk.bin). This file can be generated using the resources/tools/rsvk.sh script.}

\item{vkrs_file}{Name and path of the binary file containing the VariantKey to rsID mapping (vkrs.bin). This file can be generated using the resources/tools/vkrs.sh script.}
}
\description{
Load the VariantKey support files.
This should be the first function called in order to load the support files.
}
