% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{MmapGenorefFile}
\alias{MmapGenorefFile}
\title{Memory map the specified genome reference binary file (fasta.bin).
Returns the memory mapped file object, the columns object and the number of rows.}
\usage{
MmapGenorefFile(file)
}
\arguments{
\item{file}{Path to the file to map.}
}
\description{
Memory map the specified genome reference binary file (fasta.bin).
Returns the memory mapped file object, the columns object and the number of rows.
}
