% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{MmapNRVKFile}
\alias{MmapNRVKFile}
\title{Memory map the specified NRVK binary file (nrvk.bin).
Returns the memory mapped file object, the columns object and the number of rows.}
\usage{
MmapNRVKFile(file)
}
\arguments{
\item{file}{Path to the file to map.}
}
\description{
Memory map the specified NRVK binary file (nrvk.bin).
Returns the memory mapped file object, the columns object and the number of rows.
}
