% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{MunmapBinfile}
\alias{MunmapBinfile}
\title{Unmap and close the memory-mapped file.
On success, munmap() returns 0, on failure -1.}
\usage{
MunmapBinfile(src, fd, size)
}
\arguments{
\item{mf}{Descriptor of memory-mapped file.}
}
\description{
Unmap and close the memory-mapped file.
On success, munmap() returns 0, on failure -1.
}
