% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{NormalizeVariant}
\alias{NormalizeVariant}
\title{Normalize a variant.
Flip alleles if required and apply the normalization algorithm described at:
https://genome.sph.umich.edu/wiki/Variant_Normalization
Return Positive bitmask number in case of success, negative number in case of error.
When positive, each bit has a different meaning when set, has defined by the NORM_* defines:
  * bit 0:
      Reference allele is inconsistent with the genome reference (i.e. when contains nucleotide letters other than A, C, G and T).
  * bit 1:
      Alleles have been swapped.
  * bit 2:
      Alleles nucleotides have been flipped (each nucleotide have been replaced with its complement).
  * bit 3:
      Alleles have been left extended.
  * bit 4:
      Alleles have been right trimmed.
  * bit 5:
      Alleles have been left trimmed.}
\usage{
NormalizeVariant(src, idx, chrom, pos, ref, alt)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the genome reference data (fasta.bin).}

\item{idx}{Index of sequences offset by chromosome number (1 to 25).}

\item{chrom}{Chromosome encoded number.}

\item{pos}{Position. The reference position, with the first base having position 0.}

\item{ref}{Reference allele. String containing a sequence of nucleotide letters.}

\item{alt}{Alternate non-reference allele string.}
}
\description{
Normalize a variant.
Flip alleles if required and apply the normalization algorithm described at:
https://genome.sph.umich.edu/wiki/Variant_Normalization
Return Positive bitmask number in case of success, negative number in case of error.
When positive, each bit has a different meaning when set, has defined by the NORM_* defines:
  * bit 0:
      Reference allele is inconsistent with the genome reference (i.e. when contains nucleotide letters other than A, C, G and T).
  * bit 1:
      Alleles have been swapped.
  * bit 2:
      Alleles nucleotides have been flipped (each nucleotide have been replaced with its complement).
  * bit 3:
      Alleles have been left extended.
  * bit 4:
      Alleles have been right trimmed.
  * bit 5:
      Alleles have been left trimmed.
}
