% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ParseRegionKeyHex}
\alias{ParseRegionKeyHex}
\title{Parses a RegionKey hexadecimal string and returns the code.}
\usage{
ParseRegionKeyHex(hex)
}
\arguments{
\item{hex}{RegionKey hexadecimal string (it must contain 16 hexadecimal characters).}
}
\description{
Parses a RegionKey hexadecimal string and returns the code.
}
