% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ParseVariantkeyHex}
\alias{ParseVariantkeyHex}
\title{Parses a VariantKey hexadecimal string and returns the code.}
\usage{
ParseVariantkeyHex(vs)
}
\arguments{
\item{vs}{VariantKey hexadecimal string (it must contain 16 hexadecimal characters).}
}
\description{
Parses a VariantKey hexadecimal string and returns the code.
}
