% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{RegionKey}
\alias{RegionKey}
\title{Returns a 64 bit regionkey based on CHROM, START POS (0-based), END POS and STRAND.}
\usage{
RegionKey(chrom, startpos, endpos, strand)
}
\arguments{
\item{chrom}{Chromosome. An identifier from the reference genome, no white-space or leading zeros permitted.}

\item{startpos}{Start position (zero based).}

\item{endpos}{End position (startpos + region_length).}

\item{strand}{Strand direction (-1, 0, +1)}
}
\description{
Returns a 64 bit regionkey based on CHROM, START POS (0-based), END POS and STRAND.
}
