% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{RegionKeyHex}
\alias{RegionKeyHex}
\title{Returns RegionKey hexadecimal string (16 characters).}
\usage{
RegionKeyHex(vk)
}
\arguments{
\item{vk}{RegiontKey code.}
}
\description{
Returns RegionKey hexadecimal string (16 characters).
}
