% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ReverseRegionKey}
\alias{ReverseRegionKey}
\title{Reverse a RegionKey code and returns the normalized components as regionkey_rev_t structure.}
\usage{
ReverseRegionKey(rk)
}
\arguments{
\item{rk}{RegionKey code.}
}
\description{
Reverse a RegionKey code and returns the normalized components as regionkey_rev_t structure.
}
