% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ReverseVariantKey}
\alias{ReverseVariantKey}
\title{Decode a VariantKey code and returns the components.}
\usage{
ReverseVariantKey(hexcode)
}
\arguments{
\item{hexcode}{VariantKey hexadecimal string (16 hexadecimal characters).}
}
\description{
Decode a VariantKey code and returns the components.
}
