% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ReverseVariantKey}
\alias{ReverseVariantKey}
\title{Reverse a VariantKey code and returns the normalized components.}
\usage{
ReverseVariantKey(mc, vk)
}
\arguments{
\item{mc}{Memory-mapped columns object as retured by MmapNRVKfile.}

\item{vk}{VariantKey code.}
}
\description{
Reverse a VariantKey code and returns the normalized components.
}
