% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ReverseVariantkey}
\alias{ReverseVariantkey}
\title{Reverse a VariantKey code and returns the normalized components.}
\usage{
ReverseVariantkey(src, last, vk)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the VariantKey to REF+ALT lookup table (vknr.bin).}

\item{last}{Number of variants in the src file -1. Set this to 0 to skip the lookup table.}

\item{vk}{VariantKey code.}
}
\description{
Reverse a VariantKey code and returns the normalized components.
}
