% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{ReverseVariantKey}
\alias{ReverseVariantKey}
\title{Reverse a VariantKey code and returns the normalized components.}
\usage{
ReverseVariantKey(vk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey code.}

\item{mc}{Memory-mapped columns object as retured by MmapNRVKfile.}
}
\description{
Reverse a VariantKey code and returns the normalized components.
}
