% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{sort.uint64}
\alias{sort.uint64}
\title{Sorting or Ordering uint64 Vectors.}
\usage{
\method{sort}{uint64}(x, ...)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Sorting or Ordering uint64 Vectors.
}
