% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{sort.uint64}
\alias{sort.uint64}
\title{Sorts a uint64 vector in ascending order.}
\usage{
\method{sort}{uint64}(x, ...)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Sorts a uint64 vector in ascending order.
}
