% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{str.uint64}
\alias{str.uint64}
\title{Compactly Display the Structure of an uint64 Object.}
\usage{
\method{str}{uint64}(object, vec.len = strO$vec.len, give.head = TRUE,
  give.length = give.head, ...)
}
\description{
Compactly Display the Structure of an uint64 Object.
}
