% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{uint64ToHex}
\alias{uint64ToHex}
\title{Coerce uint64 vector to hexadecimal character.}
\usage{
uint64ToHex(x, ...)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Coerce uint64 vector to hexadecimal character.
}
