% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{unique.uint64}
\alias{unique.uint64}
\title{Eliminates all but the first element from every consecutive group of equal values.}
\usage{
\method{unique}{uint64}(x, ...)
}
\arguments{
\item{x}{uint64 vector}
}
\description{
Eliminates all but the first element from every consecutive group of equal values.
}
