% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VariantKey}
\alias{VariantKey}
\title{Returns the hexadecimal representation of a 64 bit variant key based on CHROM, POS (0-base), REF, ALT.}
\usage{
VariantKey(chrom, pos, ref, alt)
}
\arguments{
\item{chrom}{Chromosome. An identifier from the reference genome, no white-space or leading zeros permitted.}

\item{pos}{Position. The reference position, with the first base having position 0.}

\item{ref}{Reference allele. String containing a sequence of nucleotide letters.}

\item{alt}{Alternate non-reference allele string.}
}
\description{
Returns the hexadecimal representation of a 64 bit variant key based on CHROM, POS (0-base), REF, ALT.
}
