% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VariantkeyHex}
\alias{VariantkeyHex}
\title{Returns VariantKey hexadecimal string (16 characters).
Upon successful return, these function returns the number of characters processed (excluding the null byte used to end output to strings).
If the buffer size is not sufficient, then the return value is the number of characters required for
buffer string, including the terminating null byte.}
\usage{
VariantkeyHex(vk)
}
\arguments{
\item{vk}{VariantKey code.}

\item{str}{String buffer to be returned (it must be sized 17 bytes at least).}
}
\description{
Returns VariantKey hexadecimal string (16 characters).
Upon successful return, these function returns the number of characters processed (excluding the null byte used to end output to strings).
If the buffer size is not sufficient, then the return value is the number of characters required for
buffer string, including the terminating null byte.
}
