% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VariantKeyHex}
\alias{VariantKeyHex}
\title{Returns VariantKey hexadecimal string (16 characters).}
\usage{
VariantKeyHex(vk)
}
\arguments{
\item{vk}{VariantKey code.}
}
\description{
Returns VariantKey hexadecimal string (16 characters).
}
