% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VariantToRegionkey}
\alias{VariantToRegionkey}
\title{Get RegionKey from VariantKey.}
\usage{
VariantToRegionkey(vk, mc = vk.env$nrvk_$MC)
}
\arguments{
\item{vk}{VariantKey code.}

\item{mc}{Memory-mapped columns object as retured by MmapNRVKfile.}
}
\description{
Get RegionKey from VariantKey.
}
