% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VariantToRegionkey}
\alias{VariantToRegionkey}
\title{Get RegionKey from VariantKey.}
\usage{
VariantToRegionkey(mc, vk)
}
\arguments{
\item{mc}{Memory-mapped columns object as retured by MmapNRVKfile.}

\item{vk}{VariantKey code.}
}
\description{
Get RegionKey from VariantKey.
}
