% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VariantToRegionkey}
\alias{VariantToRegionkey}
\title{Get RegionKey from VariantKey.}
\usage{
VariantToRegionkey(src, last, vk)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the VariantKey to REF+ALT lookup table (nrvk.bin).}

\item{last}{Number of variants in the src file -1.}

\item{vk}{VariantKey code.}
}
\description{
Get RegionKey from VariantKey.
}
