% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VknrBinToTsv}
\alias{VknrBinToTsv}
\title{Convert a vrnr.bin file to a simple TSV.
Return Number of written bytes or 0 in case of error.
For the reverse operation see the resources/tools/nrvk.sh script.}
\usage{
VknrBinToTsv(mc, tsvfile)
}
\arguments{
\item{mc}{Memory-mapped columns object as retured by MmapNRVKfile.}

\item{tsvfile}{Output tsv file name. Note that existing files will be replaced.}
}
\description{
Convert a vrnr.bin file to a simple TSV.
Return Number of written bytes or 0 in case of error.
For the reverse operation see the resources/tools/nrvk.sh script.
}
