% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantkey.R
\name{VknrBinToTsv}
\alias{VknrBinToTsv}
\title{Convert a vrnr.bin file to a simple TSV.
Return Number of written bytes or 0 in case of error.
For the reverse operation see the resources/tools/vknr.sh script.}
\usage{
VknrBinToTsv(src, last, tsvfile)
}
\arguments{
\item{src}{Address of the memory mapped input file containing the VariantKey to REF+ALT lookup table (vknr.bin).}

\item{last}{Number of variants in the src file -1. Set this to 0 to skip the lookup table.}

\item{tsvfile}{Output tsv file name. Note that existing files will be replaced.}
}
\description{
Convert a vrnr.bin file to a simple TSV.
Return Number of written bytes or 0 in case of error.
For the reverse operation see the resources/tools/vknr.sh script.
}
